Add files via upload
Users could provide three .txt files, one test_data file is the test file which has several matrixes made by 0 and 1 (formatted such as 9*7, 8*10). The other two (training_data and target_data) is used to compare to these matrixes and the machine could recognize is the matrix represent a letter or a number or nothing.