Add files via upload
We intend to create a database that include all movies released from 2007 to 2011. The movies are classified by 5 criterias, which are Genre, CriticalRating, AudienceRating, ReleaseDate, and Budget. Public users are able to search movies by one or more criterias and registered users are able to add, delete, or modify the movie database. Registered users need to signup with username, password, and email address, their information and behaviors will stored in user table to help us analyze client preference in the future.