Add files via upload
Wechat is a mobile app that allows users to chat with each other, share their photos or article in a public space, find new friends by shaking their devices or finding others near you and users can create a group then chat in it. The app also provide the function to chat with others face to face or just like making a call.