Add files via upload
The purpose of this document is to describe the Golden Springs Realty IT system that we expect to put in place in order to alleviate many challenges the company is facing. The system will incorporate a number of subsystems that will all communicate and interact in order to make transactions and sales more efficient for tenants, customers and sales representatives.